requirements.txt
Flask==2.2.2
tensorflow==2.13.0
google-cloud-firestore==2.10.0
google-cloud-storage==2.11.0
firebase-admin==6.0.1
numpy==1.23.4
Pillow==9.2.0
opencv-python-headless==4.7.0.72
requests==2.28.1